+++ types/src/commonMain/kotlin/kotools/types/collections/NotEmptyList.kt
@@ -6,9 +6,10 @@ import kotlinx.serialization.builtins.ListSerializer
 import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
+import kotools.shared.Project.Types
+import kotools.shared.SinceKotools
+import kotools.shared.StabilityLevel
 import kotools.types.Package
-import kotools.types.SinceKotoolsTypes
-import kotools.types.StabilityLevel
 
 // ---------- Builders ----------
 
@@ -16,7 +17,7 @@ import kotools.types.StabilityLevel
  * Creates a [NotEmptyList] starting with a [head] and containing all the
  * elements of the optional [tail].
  */
-@SinceKotoolsTypes("3.0")
+@SinceKotools(Types, "3.0")
 public fun <E> notEmptyListOf(head: E, vararg tail: E): NotEmptyList<E> {
     val list: List<E> = tail.toList()
     return notEmptyListOf(head, list)
@@ -38,7 +39,7 @@ private fun <E> notEmptyListOf(head: E, tail: List<E>): NotEmptyList<E> {
         "${Package.collections}.toNotEmptyListOrThrow"
     )
 )
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 @Throws(IllegalArgumentException::class)
 public fun <E> Array<E>.toNotEmptyList(): NotEmptyList<E> =
     toNotEmptyListOrThrow()
@@ -54,7 +55,7 @@ public fun <E> Array<E>.toNotEmptyList(): NotEmptyList<E> =
         "${Package.collections}.toNotEmptyListOrThrow"
     )
 )
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 @Throws(IllegalArgumentException::class)
 public fun <E> Collection<E>.toNotEmptyList(): NotEmptyList<E> =
     toNotEmptyListOrThrow()
@@ -64,7 +65,7 @@ public fun <E> Collection<E>.toNotEmptyList(): NotEmptyList<E> =
  * returns the result of calling the [defaultValue] function if this array is
  * empty.
  */
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 public inline infix fun <E> Array<E>.toNotEmptyListOrElse(
     defaultValue: (Array<E>) -> NotEmptyList<E>
 ): NotEmptyList<E> = toNotEmptyListOrNull() ?: defaultValue(this)
@@ -74,7 +75,7 @@ public inline infix fun <E> Array<E>.toNotEmptyListOrElse(
  * returns the result of calling the [defaultValue] function if this collection
  * is empty.
  */
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 public inline infix fun <E> Collection<E>.toNotEmptyListOrElse(
     defaultValue: (Collection<E>) -> NotEmptyList<E>
 ): NotEmptyList<E> = toNotEmptyListOrNull() ?: defaultValue(this)
@@ -83,7 +84,7 @@ public inline infix fun <E> Collection<E>.toNotEmptyListOrElse(
  * Returns a [NotEmptyList] containing all the elements of this array, or
  * returns `null` if this array is empty.
  */
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 public fun <E> Array<E>.toNotEmptyListOrNull(): NotEmptyList<E>? = toList()
     .toNotEmptyListOrNull()
 
@@ -91,7 +92,7 @@ public fun <E> Array<E>.toNotEmptyListOrNull(): NotEmptyList<E>? = toList()
  * Returns a [NotEmptyList] containing all the elements of this collection, or
  * returns `null` if this collection is empty.
  */
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 public fun <E> Collection<E>.toNotEmptyListOrNull(): NotEmptyList<E>? =
     takeIf(Collection<E>::isNotEmpty)
         ?.toList()
@@ -102,7 +103,7 @@ public fun <E> Collection<E>.toNotEmptyListOrNull(): NotEmptyList<E>? =
  * Returns a [NotEmptyList] containing all the elements of this array, or throws
  * an [IllegalArgumentException] if this array is empty.
  */
-@SinceKotoolsTypes("3.2", StabilityLevel.Alpha)
+@SinceKotools(Types, "3.2", StabilityLevel.Alpha)
 @Throws(IllegalArgumentException::class)
 public fun <E> Array<E>.toNotEmptyListOrThrow(): NotEmptyList<E> =
     toNotEmptyListOrNull()
@@ -112,7 +113,7 @@ public fun <E> Array<E>.toNotEmptyListOrThrow(): NotEmptyList<E> =
  * Returns a [NotEmptyList] containing all the elements of this collection, or
  * throws an [IllegalArgumentException] if this collection is empty.
  */
-@SinceKotoolsTypes("3.2", StabilityLevel.Alpha)
+@SinceKotools(Types, "3.2", StabilityLevel.Alpha)
 @Throws(IllegalArgumentException::class)
 public fun <E> Collection<E>.toNotEmptyListOrThrow(): NotEmptyList<E> =
     toNotEmptyListOrNull() ?: throw IllegalArgumentException(
@@ -125,7 +126,7 @@ public fun <E> Collection<E>.toNotEmptyListOrThrow(): NotEmptyList<E> =
  * @param E The type of elements contained in this list.
  */
 @Serializable(NotEmptyListSerializer::class)
-@SinceKotoolsTypes("1.3")
+@SinceKotools(Types, "1.3")
 public sealed interface NotEmptyList<out E> : NotEmptyCollection<E>, List<E>
 
 private class NotEmptyListImplementation<out E>(private val list: List<E>) :